Changed type conversion from UInt8 to UInt64 in secondsToDisplay to allow for more cases.

+++ Rest Less/Rest Less/SecondsToDisplay.swift
@@ -12,13 +12,13 @@ func secondsToDisplay(var time: NSTimeInterval) -> String{
     if time <= 0 {
         return "00:00:00"
     }
-    let minutes = UInt8(time / 60.0)
+    let minutes = UInt64(time / 60.0)
     time -= (NSTimeInterval(minutes) * 60)
     
-    let seconds = UInt8(time)
+    let seconds = UInt64(time)
     time -= NSTimeInterval(seconds)
     
-    let fraction = UInt8(time * 100)
+    let fraction = UInt64(time * 100)
     
     let strMinutes = minutes > 9 ? String(minutes):"0" + String(minutes)
     let strSeconds = seconds > 9 ? String(seconds):"0" + String(seconds)
